Bug fixed again
